// src/openapi/models/PagingPlaylistObject.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* eslint-disable */

import type { PagingObject } from './PagingObject';
import type { SimplifiedPlaylistObject } from './SimplifiedPlaylistObject';

export type PagingPlaylistObject = (PagingObject & {
items?: Array<SimplifiedPlaylistObject>;
});