// src/openapi/models/PagingPlaylistTrackObject.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* eslint-disable */

import type { PagingObject } from './PagingObject';
import type { PlaylistTrackObject } from './PlaylistTrackObject';

export type PagingPlaylistTrackObject = (PagingObject & {
items?: Array<PlaylistTrackObject>;
});